ConfigTree Daemon crash during app remove/start test
When Starting a application manually and removing this app at
the same time, the system will crash since config tree of
the app is removed but supervisor still bring up the app with
empty config tree.
The solution is to detect the scenario when App is going to be
removed or updated, reject the app start of this App directly.

Resolves: LE-13300
Related: LEAR7M-686
Change-Id: Iebbcb7c7235838c55d138667a37b6fed06bba96d