Use prefix length for le_net_ChangeRoute
The 3rd parameter of le_net_ChangeRoute expects:
- in case of IPv4, a netmask (255.255.255.0)
- in case of IPv6, a prefix length (32)

To make things uniform and align with iproute2, which only takes
prefix length, change the third parameter.

As to avoid this making a breaking change, the parameter still
accept a netmask and converts it to a prefix length.
This is to maintain compatibility only. The form with netmask is
deprecated and the compatibility code will be removed in a future
release.

Resolves: LE-12820
Change-Id: Id92f8e30cbc782c3c29456f472a3103da2f643a6