Revert "Merge mkTools changes since 18.09.0.rc1 merge into uLegato."
This reverts commit 53d7acb11b82110633603ab22d9bf7d35c61e363.

The reverted commit is causing legato to enter a watchdog triggered
boot loop when it is built against the pre-built PA libraries.  The
reason for this is still under investigation, but the change is being
reverted to unblock the 19.07 release.  Reverting this patch may
adversely affect RTOS (FreeRTOS and ThreadX) from Legato master.

Resolves: LE-13204
Change-Id: I002597bafe4d352ebb4bb1bc254692a5ada6c1b2