[mkedit] Make creating missing directory an editable action
If app, module search directory path listed in the
corresponding search sections in sdef does not exist then create the
directory before attempting to add any file inside the directory.

This is being added as an edit action which is undo-able in case
of errors. If the definition file's containing directory did not
initially exist and an error exception occurred then the
containing directory will be deleted if created.

Miscellaneous:
- Add a function to print logging for verbose or dry-run option.

Resolves: LE-13237
Change-Id: If0285ef97e7bc3da6db5baa511edb421bd217477