Missing preloaded app leaves the system in a bad state
In a case of preloaded app, the system is unpacked and installed
with a symlink that may be broken if the app was not in fact
pre-installed. Currently, such app cannot be removed or re-installed
because CopyRecursive() does not tolerate broken symlinks.

The fix is to make file_CopyRecursive() able to handle a symlink,
whether it's broken or not.

Resolves: LE-11481
Change-Id: I1f1b907056a715c60adbe56c9df066ae80e1998b