Add build vars to exclude libs
Inclusion of zlib, curl and openSSL libraries into apps
causes a massive increase in the size of the legato.cwe,
preventing it from fitting on some devices when combined
with apps, such as the mangOH apps and Octave edge apps.

Three variables are added to help reduce the size by
excluding libs from the built legato system.  This is
possible because those libs are also available in the
root file system.  The variables are:
  USE_ROOTFS_OPENSSL
  USE_ROOTFS_CURL
  USE_ROOTFS_ZLIB
Set them to 1 to cause exclusion.

Using the rootfs lib has some downsides in term of compatibility
between what the software was built with (toolchain)
and the libraries present on the rootfs.
Since this is an opt-in change, it is assumed that people
taking libraries from the rootfs understand that risk.

Resolves: LE-13277
Change-Id: I35aac284442c888cdcff66cda0300cf2569b34f0