Fix for "secstore readmeta" command not working.
The problem is that in new TrustZone-based implementation, the meta
file does not exist, and can't be simply copied to the provided path.

The fix is to generate the meta data on demand, and write it to the
provided file path.

Resolves: LE-13343
Change-Id: I80791fa7aa790369be9e9eaf20de45a3605dd8f0
(cherry picked from commit da17c310d4452117ac3534a91ecab36744560ddd)