Port le_hex, le_pack, and le_pathIter components
We not only port these components, but also change the corresponding
tests from the old CMake/bash style framework to the new adef/cdef style
framework.

The trickiest part is to remember calling pathIter_init() in
InitLegatoFramework() in microSupervisor.c, or you won't be able to use
the memory pool.

Another important design decision in this commit is putting all
cross-target initialization in InitLegatoFramework() in
microSupervisor.c. For example, we call setlocale() here to initialize
the standard C locale system, so that hex.c can call locale-dependent
functions, such as toupper().

Resolves: LE-13076
Change-Id: If8c520a3e887510f73a2df4329b0a33aa0f5578f