Update legato-qemu to support wp76xx/ar758x
Rework target files to use to type of Dockerfile
  - (alpine) uses official qemu package from Alpine distribution
  - (build) uses Sierra fork of QEmu to support Sierra targets

The qemu fork is extracted in legato/3rdParty/qemu/ (currently v2.6.2)
and is picked up by build-qemu.sh (and Dockerfile.build) to build it.

legato-qemu is updated to handle info provided along the kernel & rootfs
(packaged in a 'rawflash' file) as a JSON file.
The new format is a little more flexible than the original qemu-config
format that was just a list of VAR=value declarations.

Resolves: LE-13386
Related: LXSWIREF-345
Change-Id: I8ee8358ce19cc4acadd33e08ad694af85f899411