[MCU] Asynchronous Connection Process Takes Too Long
These changes address an issue affecting the re-connection
sequence and its dependency on the FD Monitor callback.
Under certain circumstances, new data may be arrive and be
waiting in the receive buffer of the serial UART,
but FD Monitor will not trigger a new callback, thereby resulting
in a 'stuck' start-up sequence.

Resolves: LE-13377
Change-Id: I111b44a63fa219a1dc660d99193a1d4f0449e78d