Fix for failure sending non-ASCII SMS
SMS encoding can be anything, not limited to UTF8. To copy the
SMS text, "le_utf8_Copy()" is used. This routine may fail to
copy non-UTF8 text, because it is examining the upper bit to
determine number of bytes in character. Therefore, the text
that is encoded differently, may not be copied correctly.

The fix is to revert the LE-12273 change and use "strncpy()",
which is agonostic to the encoding, plus ensure that
the destination is always NUL-terminated.

Resolves: LE-13407
Change-Id: Iab30d672ff683376eb41ed5c16f115bd36d00587