Add pytest to the HL76 ThreadX platform
To enable pytest framework, we need three key modifications. First, we
modify the build script to embed the frameworkAdaptor repository path in
the generated config.sh script, so legato/conftest.py script can get the
path of frameworkAdaptor.

Second, different targets have different CLI implementation, so we place
the platform-dependent implementation in an adaptor Python module
located in the frameworkAdaptor root directory. When we run pytest in the
legato repository, pytest knows where to load the adaptor module and run
the platform-dependent logic.

Third, we add a reset CLI command, so we can reboot the target to a
known state at the beginning of each test case.

This implementation cannot run the frameworkAdaptor test cases under the
frameworkAdaptor repository, because pytest can only collect and run
test cases under the root directory (legato). This issue will resolved
in another ticket.

Resolves: LE-13100
Change-Id: Icbca57399e94ed9f79ff11e5ee04fb60ab806356