Fix for WatchDog not working for test app "monitoredService"
The problem is, that if the watchdog is built using "mkapp" for
a Linux target (e.g. wp76xx), then the config variable
LE_CONFIG_LINUX may not be imported, and as a result le_wdog.api
is included as "optional". This results in the failure of the
TryConnectService() call, and watchdog not being activated.
The fix is to invert the conditional compilation logic so that
even if the LE_CONFIG_LINUX is not set, the app still would include
the le_wdog.api correctly.

Resolves: LE-13515
Change-Id: Ia70d9083c9e51256d63598850629199793a17efe
(cherry picked from commit 715c229e18b084a1f3ae49ea66e0c94cf5f2e66a)