Fix for number of supplementary groups not being initialized.
If the app is sandboxed and there are no groups configured, then
the data member numSupplementGids is used without being initialized.
The fix is to memset to 0 the whole newly allocated structure
(to avoid similar problems with other data fields) and to reset
it to 0 if no groups are found in the config tree.

Resolves: LE-13583
Change-Id: Ia9fa367ecde58198080abab2909c8cb09f1ceff9