Fix for GPIO pulsating on repeated Activate calls
The problem is when "out" is written to the direction file in the sysfs,
the Linux kernel will overwrite the current value of the GPIO
and set it low. Legato will then continue to set the GPIO high.
The net result is that there is a very brief low pulse on the GPIO if
it is already set as an output, and is set high.
The fix is to skip writing the direction to the file, if the direction
is already correct.

Resolves: LE-13244
Change-Id: If408e5af608bcc255c76a07d2bcc232a383f42a2
(cherry picked from commit 5e715d75f5fab1ebf00ebe29089625da6ad24a4c)