Port modemServices to HL76 on default PA
Some compilation errors are fixed and features are aligned with
HL78 to build modemSerivces on the default platform adaptor.

This is version 2 of this patch as the first version broke modemServices
on Linux.  Adjusting the initialization order to be more similar
to the order on Linux appears to have fixed that problem.

Resolves: LE-13484
Change-Id: I48db8f5d83ec81074117e241f2f0bf878bdce0a2