Fix le_data's channelRef passing in auto-adding host routes for DNS servers
When le_data's config specifies to not add a default route to the default GW
address for the data channel that it has established, the wrong channelRef
passing in SetDnsRoutes() caused failure in such auto host route adding.
This fix changes the internal function le_dcsTech_GetDNSAddresses() to pass
DCS's channelRef than the tech's channelRef to correct it.

Resolves: LE-13745
Related: LE-13725
Change-Id: Ia45d59747a485ecabc4792c28bbf6ad385242124