Return stack sizes to 8K
modemService stack was increased in an attempt to fix LE-13758.  With
a final fix for this issue in swiQmi, this stack can be returned to
a saner size of 8K.

To decrease porting effort and reduce memory usage, remove le_lpt
which is not used on hl78 so should not be required on gill either.

Also create a static assert macro which works pre-C11.

Resolves: LE-13758
Change-Id: I7afee723aa98c51559dc71eb4aae01a3415d6a0c