Start new thread for CLI commands.
Start each CLI command in its own thread, as a regular app processes.
This allows them to bind to APIs and perform IPC as any other task.

With this in place, modify the "reset" app to use the ulpm Reboot API
to do a clean reboot instead of the previous dirty reboot.

Resolves: LE-13555
Change-Id: Ife8481562c3f1c1b8ac34e0644664876e36be86b