[CM] Fix cm's wrong behavior upon dual stack PDP
Current implementaion of cm script will call the fucntion
update_dns for each supported IP protocol. It will back-
up the existing /etc/resolv.conf and create a new copy.
If supported PDPs are both IPv4 and IPv6, this function
is called twice. The second time it is called (for IPv6),
it wipes the one previously created (for IPv4). This
leads to a /etc/resolv.conf contains only DNS for IPv6
in this case.

Now moving the call to the function out of the loop, and
passing a list of all DNS's could resolve the issue.

Resolves: LE-13807
Change-Id: I418afbf43bba050f51e9c762b8b4dfe1036f586d