Provide a workaround for fdMonitor issues
fdMonitor may fire POLLIN events at the wrong time when we signal the
event too fast. This issue is described in LE-13894.

We provide a workaround for this blocking issue by adding a tiny delay
in the tight loop where we try to send queued unsolicited responses
after sending the final response. When we resolve LE-13894, we should be
able to remove this workaround

Resolves: LE-13896
Change-Id: Id5bfde2549a60b368291d239572e280e67bb66be