Register commands added before atFwdLinker starts
In atFwdLinker, we add a command registration handler, which registers
AT commands to the modem firmware. However, apps that start before
atFwdLinker may have added AT commands, which are not handled by the
command registration handler.

To solve this race condition, we call the command registration handler
for these already existed AT commands when we call the
le_atServer_AddCmdRegistrationHandler function to add the handler.

Resolves: LE-14036
Change-Id: Ica0dd199b9c19967c5282c258e0765c5d3c46ecb