Inline portions of client stubs
The client stubs have a very small function to add the remote connection
reference, then a larger common code to handle pack/unpack.  Inline the
small stub, as the function call overhead for this function was a
significant portion of the code.

This change saves 16KB of code space on rc76.

Resolves: LE-14044
Change-Id: Iaffa9da1e7b03390fcaad46c536f6d53dbe320d7