[fwupdate] Fix fwupdate app fails to responds to host command
The essence is that the permissions were granted to "root",
and we changed it to "powerMgr". The fix is about granting
permissions to subject with SMACK label "app.powerMgr".

The definition file is used to create the "bindings" section
in the config tree, which in turn is used to give the
permission. So binding to "root" or "powerMgr" is equivalent
from the execution viewpoint, but produces different results
in terms of permission.

Resolves: LE-14080
Related: LE-4096
Change-Id: I8f47235493ae171025d574097a18c7b58ccf739b