Change legato view to remain visible upon error
Previously when an error occured the legato view would be entirely
hidden. This provides a bad user experience and offers no opportunity to
resolve any errors from within the legato view.

Now the legato view remains visible and allows the user to fix the error
that is displayed to them in a notification.

Resolves: LETOOLS-1034
Change-Id: Icc0e2594795d8ff088286739754327d2a269d6bb