Follow HTTP/1.1 request standard
The HTTP/1.1 request standard requires an empty line at the end of the
request header, but we didn't follow the standard. That's why we always
get "HTTP/1.1 400 Bad Request" response.

By following the standard and adding an empty line at the end of the
request header, we can get expected responses from any web server.

Resolves: LE-14008 LE-14234
Change-Id: Ib01fed0f5212dc9ecb365e444d0cb12382b14f41