Fix non blocking ConnectService using local messaging
If api requirements are specified with [manual-start] keyword the automatic call to ConnectService
will not be generated. In this case, the client may want to use TryConnectService for non blocking
connection attempt to the server. Due to a small error in msgLocal_TryOpenSessionSync, this
mechanism was broken and could result in client mistakenly thinking they are connected.

Relates: LE-14348
Resolves: LE-15896
Change-Id: I173dfe0d98597a38b2051cf1af0567a3c8bc2a17