Keep receiving the body for POST/PUT/DELETE
When the Content-Length field in the HTTP response header is larger than
zero, we should keep trying to receive the data, until the entire body
specified by the Content-Length field is received.

We do have this logic for the GET command, but we forgot to apply the
same logic to the POST, PUT, and DELETE commands. That's why we may not
be able to get the response body when we submit a POST command.

This patch can enhance the original logic, so we won't miss any HTTP
response data.

Resolves: LE-14373
Change-Id: Id77556ae70ac0527a6c390af14c9032547aa9820