Fix how we add items to hashmap
When we insert a new item whose key already exists in the hashmap, we
don't update the pointer to the new key but keep the pointer to the old
key, assuming that the old key still exists after releasing the old
value memory block. It will cause trouble when both key and value are
in the same memory block.

For example, in socketMonitor you can see the following code snippet:

oldPtr = le_hashmap_Put(MonitoredSockets, &addPtr->fd, addPtr);
if (oldPtr)
{
    le_mem_Release(oldPtr);
}

As you can see that, the key (addPtr->fd) is a field in the value data
structure (addPtr): they are in the same memory block. When we insert a
new value memory block (addPtr), we still keep using &oldPtr->fd. Since
we release the memory block pointed by oldPtr, &oldPtr->fd is not longer
valid. All kinds of strange issues start from here, including memory
leak.

To resolve the issue, we should always update the hashmap item so it
always has a valid pointer to the key memory block.

Resolves: LE-14511
Related: LE-14052
Change-Id: I3dbd7342683038600f70839f0e7f7da0b3a03f27