Fix inspect timers with multiple threads
When there are no more thread objects on the list (or list is empty),
TimerTypeIndex = TIMER_WAKEUP, to read all the wakeup timers of the previous
threads, need to retrieve the address of the first thread in the list,
read from memory and update the current thread. Then take the entire
timer value in the current thread.

Resolves: LE-14445
Change-Id: I7f80a7e1cf61198bad04a38be75709316f78cdb3