[cherry-pick] Improve priority mapping in RTOS environments
Previously, the default priority always mapped to the lowest, which
made it impossible to specify a task as having a priority lower than
the default without explicitly specifying the priority of every other
task.

Now, when only realtime priorities are on, map the regular priorities
to relative realtime priorties to ensure there is some wiggle room for
those who are explicitly configuring rtX values.

Resolves: ATSWI-90
Related: LE-14459
Change-Id: I4e4ef87ac5f3b13687bd0ef07165e295940a9c45