Enlarge mem pools to support 32 HTTPS connections
RC76 inherits the behavior from HL78, which only allows two simultaneous
HTTP/HTTPS connections. However, since we claim that RC76 supports 32
HTTP/HTTPS sessions, it is very confusing and hard to explain to our
customers that why not all the sessions can establish connection.

To resovle this confusion, we want to ensure that all 32 HTTP/HTTPS
sessions can establish connection. To achieve this goal, we need to
enlarge two memory pools, HttpContextPool and SocketPool. In this
commit, we focus on increasing the size of HttpContextPool according to
the maximum number of HTTP/HTTPS sessions allowed.

Resolves: LE-14784
Change-Id: I54bff48e0797d2a9691097cee919bc93fc074a05