RPC Proxy File Stream Feature
This feature adds support for the "file" type in .API files over rpc
Proxy.
Currently only read only and write only file descriptors are supported.

Wherever an ipc message is being converted to an rpc message, it is scanned for any file descriptor
to send. If a file descriptor is found, a file stream instance is created to represent and hold that
file descriptor and related information like direction. The file Stream id and its flags are then
embedded into rpc message being generated.
On the receiving side, whenever an rpc message is received, it is scanned for any stream id, if a
valid filestream id and its flag are found, a filestream instance is created based on that. After
this step, both sides will have their corresponding file stream instances. Then the side with
INCOMING direction will request as much data as its local file descriptor can accept. The OUTGOING
side will then try to transmit the same amount of data if available.

Resolves: LE-14677
Change-Id: Iff2218dc3c7de2637a478caaeb1bcc547ce6f567