Add watchdogChain component requirement to missing components
The watchdogChain component must be included in the apps that may need
watchdog regardless of whether the LE_CONFIG_SERVICES_WATCHDOG config is
set or not. LE_CONFIG_SERVICES_WATCHDOG Then determines whether watchdog
is enabled or not.
There are three components that are updated: modemDaemon, dcsDaemon,
clocksync, and posDaemon.
In the case of posDaemon, when watchdogChain component is missing, a call
to le_wdogChain_Init at posDaemon COMPONENT_INIT crashes because positioning
service does not have the watchdogChain component key in its component data.

Resolves: LE-15045
Change-Id: I6891495c70b04189a583376651ae87b1afba9459