Fix the issue that allocates memory for pipe which already exists
Before this patch, OpenPipeDevice() calls le_fd_MkPipe() which
always tries to create a server side pipe first and then release
the memory if the pipe already exists. After this patch,
OpenPipeDevice() will directly open the server side pipe, because
this pipe has been already created in COMPONENT_INIT_ONCE of port
service and thus we don't have to try to create it before openning
it.

Resolves: LE-14695
Change-Id: I049b978bf498cfc5ada62b84588ab9d620cf2b7d