Fix Memory Corruption in api_Messages Pool
Problem:
RpcProxy uses le_msg_GetMaxPayloadSize to get the size of an ipc message
before repacking for send. le_msg_GetMaxPayloadSize returns the same
value for all ipc messages of an api which is the maximum size.
The rpcProxy daemon then relies on this size to check how much of the
payload it should process for repacking. It is possible that when a
client is packing parameters into a local message, it uses a buffer
previously used by another function of this api that has had more
parameters to pack. This way when the second function finishes packing
all its parameters, there is still some packed data from the first
function left at the end of buffer. Since both these ipc messages
report the same size(The maximum size), the rpcProxy repack mechanism
does not know when to stop repacking and thus during the repacking of
second function's message, it would repack more than it should.

Solution:
A change is made in the common client template code so that an
LE_PACK_EOF tag is added to the end of the message.

Resolves: LE-14992
Change-Id: Id57c0574b08c27107fe46a4b9e95901fa43a386e