[LE-15156] Authorization: Basic is automatically added to the header
Issue Description:
For a AT+KHTTPPOST transaction, the line "Authorization: Basic Og==" is always
added to the header even when the login and password credentials supplied to
AT+KHTTPCFG are empty.

Root Cause:
In the le_httpClient_SetCredentials() function in le_httpClientLib.c, the
credential field is incorrectly set to ":" by the snprintf function
when the login and password credentials are empty or NULL.
As a result, the BuildAndSendCredential() function will fail to detect
that the credential field is NULL (where it would usually return
LE_UNAVAILABLE).  This causes the creation of a Authorization: Basic
header line with the base64 encoding of the ":" colon in the degenerate case.

Solution:
Ensure that the credential field remains NULL when the supplied login and
password credential fields are empty.

Limitations:
- none

Resolves: LE-15896
Related: EURY-3183 LE-15156
Change-Id: I1cf15ed75b373b1e7ea45cd67c79802a0becbb1d
Signed-off-by: Gordon Finlay <gfinlay@sierrawireless.com>