[smack] Enable Legato SMACK for Kernel 4.14
Part 1: This is a re-commit of Jen Chitty and Bertrand Roussel's previous work
with some modification due to conflicts in source.

This commit introduces a new function to utilize SMACK's IPIN/IPOUT labels for
socket communications. Such function is used when processes want to do IPC via
socket with daemons. E.g. any process and syslogd. Making used of this new
feature resolves the socket permission errors leads to a reboot loop.

A new function for setting the SMACK64TRANSMUTE label of a directory is also
added. It may be useful for future development.

Resolves: LE-14406
Change-Id: I1ac02114ce9eeadb0d35747484fde895423b7d25