Fix Possible Memory Leak in EventPool
Fix memory leak in EventPool when the thread is terminated during
processing of the event. This happens because the event is already
removed from the event queue but code never returns from the queued
function or handler to release the event.

Solution is to store pointer to the event being processed in the
preThreadRec and release it in eventloop destructor.

Resolves: LE-15221
Change-Id: I97f25b4403699ae86862fb5a36801607dda6b8eb