[smack] [svcdir] Correct SMACK Labeling for Socket Communication with SvcDir
Service directory is acting like a messenger such that a server app would send
messages with it sits in the middle. In terms of socket communication, the
client socket is the "channel" server app and SvcDir communicate through.
Similarly, the server socket is the "channel" client app and SvcDir communicate
through. In other word,
 * client socket's SMACK64IPIN should be server's label
 * client socket's SMACK64IPOUT should be SvcDir's label
 * server socket's SMACK64IPIN should be SvcDir's label
 * server socket's SMACK64IPOUT should be client's label

Some system framework apps that need to use SvcDir should also be granted with
write access permission.

Resolves: LE-15371
Change-Id: I2af4bbac3937af455453521bdf5dbf45c11b5b76