Adds ability to initate TLS handshake on unsecured socket
Allows a user to start a connection unsecured and to initiate a
TLS handshake at a later time. This is to support explicit FTPS
where FTP commands are sent first on an unsecured channel to
determine how the data/control channel should be secured, followed
by a TLS negotiation to secure the control channel.

Also included in change:
- Updates socketTest to run on RC76

Resolves: LE-15350
Change-Id: Id4fc5d67650daec1597e384d0c110b058f127252