Let dcs_RemoveEventHandler always delete a client app's event handler upon process exit
When "cm data connect" is the trigger to start a data connection, upon its process exit,
its connection should be left to stay active, but not its event handler. This memory
corruption was caused by leaving it stay without its parent process so that an event
notification would have come to an exited process.

Resolves: LE-15367
Change-Id: Icd31fc1a02e727ec370841f1f23a19d61fadb057