Fix KConfig Default Target
Prior to this fix, the KConfig default hardware target will incorrectly
be set to localhost when using "make menuconfig_wp77xx" for the first time.
This causes important apps to be missing such as atAirVantage, atQmiLinker,
etc. This fix will correctly set the default target to wp77xx in KConfig.

Resolves: LE-15426
Change-Id: I1d1f447a03e7dba90a51598ad895fad0e3f9b042