[cherry-pick] Fix KConfig Default Target checks
Prior to this fix, the KConfig default hardware target will incorrectly
be set to localhost when using "make menuconfig_wp77xx" for the first
time.This causes important apps to be missing such as atAirVantage,
atQmiLinker,etc. This fix will correctly set the default target to
wp77xx in KConfig.

Resolves: LE-15520
Change-Id: I9d8b6719737ff829657f0781a3b0a6c31a85eabe