Fix bug in config tree's write function
If the buffer size to write is 0 while writing to config tree, the first
byte indicating the node type is being missed to write after the file creation.
This leads to node type mismatch when reading from the file, as the node type
read does not match the expected node type. This eventually leads to failing
sanity check in decrypting encrypted data in IoT Key Store.

This patch fixes the issue by allowing to write the first byte indicating
the node type to the file even though buffer size is 0 and if the buffer
pointer to write exists.

Resolves: LE-15525
Change-Id: Icbecc4e7672d57ad54aca7b137d08b2818205368