[mktools] Fix bug in generating and handling binary app package
Problem:

In summary, there were 2 issues:
1. When generating bindings in adef generator code, server type internal
was missed to handle, because of which the bindings were generated as
external bindings instead of internal.
2. In the generated adef from the binary app package, executables and
components were not recognized even though the executables were listed
as binary bundles.

Solution:
1. Add check for internal server type when generating bindings and skip
printing the app name to generate the internal bindings correctly.
2. Allow app modeller to recognize executables if they are listed as
binary bundles and allow any components within the executable.

Resolves: LE-14946
Change-Id: Ic61ff7981db6d30b3888e77f55a049aa18b01b07