Fix Inherited Pipe SMACK Labels And Update All Scripts
Currently, when the supervisor forks a new process, it contains the supervisors pipe
which is used for IPC but it also inherits the supervisor's admin label on those file
descriptors as a result.

Now, when we're in the child process, we will relabel the file descriptor SMACK labels
back to the child app's label, allowing that process to access the pipe.

Currently, all the scripts are running using python which has now been removed from the
build and thus cannot be ran. Moreoever, there are issues with where the busybox
applications being placed inside the /usr/bin rather than /bin because they're symlinks
back to /bin. Furthermore, there are issues with several missing libraries required by
some of the scripts. Also, server-upload-dirs wasn't set for where we can place files
uploaded onto the server.

Now, all the scripts are converted, however there are some issues running them because of
the strict controls from SMACK, such as legato version isn't able to run inside the app
due to privillege issues. Furthermore, running the command PS will not display all
processes due to the same reasoning above.

Note: The instapp.sh works, but there should be some checks in place to make sure it's
a valid update file. Currently, just a rough draft to make sure everything works.

Resolves: LE-15292
Change-Id: I11adf74e7e4a5e9be0d46949ae316ceacff29745