Change EID return code depending on if eSim is available
Currently, LE_FAULT followed by EXIT_FAILURE occurs when we attempt to
get the eUICC ID of a non-eSim.

Now, when we aren't using an eSim it will return LE_UNSUPPORTED and
EXIT_SUCCESS with a warning that we aren't using an eSim. Also changes
to check if SIM is present, if not we will return LE_FAULT.

Modified test case to change SIM ID to eSim prior to attempting to retrieve
eUICC ID.

Resolves: LE-15744
Change-Id: I1af7c343d80a69b9e8a15c53736ceb60a8277796