[IPC] Fix compatibility with legacy C application
In the legacy app use case, the only exposed functions to init
the library are <api>_(Try)ConnectService().
The newly introduced <api>_InitOnceCommonData(void) is not declared
in the <api>_interface.h header and even if it was, it would require
a change in the client app to make it call that function.

Instead of asking for that change, introducing a check to make sure
that the one-time init data is initialized before proceeding.
This lazy-init insure that no change is required on legacy apps.

Resolves: LE-15693
Change-Id: I12958e7b5019b8b5ddf38f4e779c6f15a444b795