Allow dynamic memory pool location.
With the introduction of overlays (or run groups), the memory usage
for memory pools may change depending on the active run group.
Instead of using fixed symbols to determine the location of memory
pools, create a new api le_mem_InitHibernation() which will accept
the actual location of memory pool memory.

Resolves: LE-15819
Change-Id: Ie0110336881f33f46e45a3e02ff67a6cb62acc1d