Print user specified versions from fwupdate query and cm info
Feature will allow users to specify any number of custom system
versions in a systemVersions section of a .sdef file, and have it be
appended to the bottom of the fwupdate query and cm info response.
The .sdef syntax is of a key-value pair, where the key is the version
name and the value is the version or path to a file containing the
version.
Test of this feature is in fwupdateUnitTest.

Resolves: LE-15609
Change-Id: I4154795147323612da4ce129bce246364680ad6e