Correct rpcProxy repack logic
The rpcProxy repack logic always placed output strings at the end of a
message instead of in the correct location within the message.  This lead
to incompatibilities when communicating between RTOS and Linux RPC proxies.
It also appears there were bugs in the code which would manifest as
corrupted output strings.

Modify the logic to ensure the packet passed over the wire always matches
the Linux "RPC" format by:

  1) Adding semantic tags to output parameters when being passed out so
     RTOS RPC proxy can know where output parameters should be placed in the
     parameter stream.
  2) Modify rpcProxy to look for these tags, and insert output parameters
     at the semantic tags when repacking the output message.

Resolves: LE-15906 LE-15447
Change-Id: Ic9c790c476771e60d92c65f6be0b6168f3c12d03