Fix FTPSND freeze
Issue Description:
Module keeps in data mode after sending data size equal to or greater
than value of <size> parameter in +KFTPSND command and <offset> = 0.

Root Cause:
A fix for an HL78 issue uses a le_fd_Read() to flush an fd. RC76
differs from HL78 in that file descriptors are non-blocking. When
the le_fd_Read() is called, it blocks and nothing else is written
to the FD.

Solution:
Make serial fd's non-blocking

Limitations:
None

Resolves: LE-15642
Change-Id: I364b68bf18764382748159131cd2a5d3d77c6c86