Align local argument handling with Linux
On Linux, output arguments where the output buffer is NULL are never
transmitted across the wire.  Their sizes are omitted from the input
parameters, and nothing is transmitted in the result.

A previous RTOS change had modified this to always transmit the output
buffer size, even if there was no output buffer.  Restore the previous
behaviour of not transmitting the output buffer size.

To align with Linux, also use the Linux behaviour of not transmitting
any information for optimized parameters (strings and byte array
buffers) if the output buffer pointer is NULL.

Resolves: LE-15992
Change-Id: I9dc29d536dcfcc36d38a797696923fbcd01e5a8a