Adding further restrictions to the AT Proxy Parser
These patches:

- Ensure that AT parameters start with either a digit (0-9),
a quote, or a comma.   All other characters will only be
accepted if within a quote (i.e. are part of a string).

- For all READ operations, confirm that there are
no additional characters between the operator ("?") and the carriage return.
If so, throw an error.

- For all TEST operations, confirm that there are
no additional characters between the operator ("=?") and the carriage return
If so, throw an error.

Resolves: ATSWI-190
Change-Id: I916edd34370a7f1d238956c4651cc83fef78367b