Increase CClient stack size to 8192 bytes
The C test client uses between 3-4K of stack.  Since some platforms
have a default stack smaller than this, manually specify the stack
size.  8K was chosen to allow for future expansion; exact stack
size is not critical as this is a test program not compiled into the
default image.

Resolves: LE-15971
Change-Id: I2c3373aaf59f0d65541d6814ad28fb1a48ff9945