Extended error code setting persistence on MCU
These changes:

- Add three new platformAdaptor functions to the pa_atProxy file:
pa_atProxy_EnableExtendedErrorCodes(),
pa_atProxy_DisableExtendedErrorCodes(),
pa_atProxy_GetExtendedErrorCodes().

For the MAP PA, these are implemented using the original logic
that relies on a static/local variable, ErrorCodeMode.

For the MCU PA, these are implemented using the new MCU filesystem
to store and persist this data in flash.

Resolves: LE-15989
Change-Id: I4d1c0cd8cd6329dc6245808e9e4760ac01bcf7c7