Enable setting cipher suites in KHTTP and KFTP
Although we can set cipher suites in KTCP, we cannot set cipher suites
in KHTTP and KFTP. The root cause is that KTCP uses the securembedtls
library for secure connection, but KHTTP and KFTP use another
secSocket_mbedtls library, which does not support setting cipher suites
as the securembedtls library does.

To fix this issue, we enhanced the secSocket_mbedtls library, so that
now it can also set cipher suites. However, the platform adaptor for
OpenSSL is not implemented in this patch.

We also added the mbedtls log mechanism in the secSocket_mbedtls
library, so we can see mbedtls log messages and verify if the cipher
suites have been set correctly during the TLS handshake process.

Resolves: ATSWI-114 ATSWI-121 LE-16017
Change-Id: I32cae9e72e499945da69887f310444eb927d1400