Clear SMACK Rules when SMACK is disabled
Currently, when SMACK is disabled via update tool, it doesn't
necessarily mean that SMACK is actually disabled, because the
rules haven't actually been cleared. This only takes effect once
the kernel has a chance to restart and check if smack is disabled.
And clear the settings from /sys/fs/smackfs

Now, we will write the default rules from /etc/smack/accesses
into the /legato/smack/load2 before we unmount /legato/smack,
and in doing so the changes are also on /sys/fs/smackfs/load2.

Resolves: LE-16015
Change-Id: I35d479d5487c42fb476fa09557ce38250f956484