Addition of "tags" section in .adef
The addition of the tags section will allow users to specify key-value
pairs by per application basis in the .adef file. This section is used
for the "disableAirVantageManagement" varaible and can be expanded with
other variables specified by the client. Tags feature is documented in
defFilesAdef.dox for external documentation. Description includes tags
feature as well as the "disableAirVantageManagement" variable.

Resolves: LE-15611
Change-Id: I61dbd67256aad7a69218b91e4a6be54c6376b645