Refactor inspect tool to be shared across targets
The inspect tool is used to debug Legato objects.  Versions of this
tool exist for both Linux and RTOS, but although they were based on
the same original source code they used different source code.

This change drops the RTOS version of inspect tool, and completes
some refactoring work which had been started in the Linux version
so the Linux version can run on both RTOS and Linux products.

Resolves: LE-16070
Change-Id: I64ddbaf9479a631420a52c37ddbe72b4a472e4d4