Add option to enable mutex & semaphore inspection on RTOS
Inspecting mutexes and semaphores was disabled on RTOS because
it requires substantial amounts of extra memory.  Also there
was no inspect tool for RTOS which supported viewing mutexes
and semaphores anyway.

Now that the Linux inspect tool has been ported to RTOS, change
the configuration option for mutexes and semaphore tracing
so it can be enabled on RTOS, but leave it disabled by default.

Resolves: LE-16070
Change-Id: I000a263546041d2f488143d73dad5f3b696974b8