Output original user input instead of IP address
When we do DNS query and successfully get the IP address, we will
overwrite the original user input with this IP address. That's why the
+KFTPCFG command cannot show the domain name input by users.

The solution is querying the IP address when we create the socket, so we
do not have to remember the IP address and overwrite the original user
input.

Resolves: ATSWI-204
Change-Id: I7708a49322a3e4f874e3248a9d546947db6788f9