Optimize AVC connection
Currently, upon every disconnection of AVC connection, the dtls connection is freed. This results in a new handshake on every new AVC connection. This patch will keep a global list of dtls connection to keep track of the timestamps to determine whether we need to perform a dtls resume with existing connections.

Resolves: LE-16035
Change-Id: I26d538fde2b05c4c110afd9ef41c6d19bbfda0d6