Don't allow '-' and '+' in numeric parameters
According to ITU-T Rec. V.250 AT command standard, a numeric value
should be a positive number, and '-' and '+' are not allowed in a
numeric constant.

The patch makes le_atServer being able to detect these invalid
characters in numeric parameters, so when users input negative numbers,
such as -1, +CME ERROR: 3 will be returned.

Resolves: ATSWI-223
Change-Id: I94b9fd2da231db9d0a31b3dbf662e9d12bfe8394