Release resources when fail to connect to a server
When the +KFTPCFG command fails to connect to a server, it does not
release resources that are not needed. In this case, the
FtpClientConnectServer function does not release the connection timer
when it fails to query the IP address of an invalid domain name.

Since the FtpClientConnectServer function uses several LE_FATAL_IF()
macros to check if resources have been released, it will crash the
system if we forget to release sessionRef->timerRef when the
FtpServerDnsQuery function fails.

We fixed this issue by properly releasing resources when we fail to
establish a connection, so that when we try to establish a connection
again in +KFTPDEL, +KFTPRCV, +KFTPSND, etc., we won't trigger the
LE_FATAL_IF() macros in the code base.

Resolves: ATSWI-229
Change-Id: I2e623e9809c2179d8fd60851a95b69ffc772a248